# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gardien API",
    "description": "API-first contract for the Gardien PWA, future mobile client, and ESP32 layer.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "def-0": {
        "description": "Standard error envelope.",
        "type": "object",
        "required": [
          "statusCode",
          "error",
          "message"
        ],
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "title": "Error"
      }
    }
  },
  "paths": {
...
    "/api/plants/": {
      "get": {
        "summary": "Search the plant directory",
        "tags": [
          "plants"
        ],
        "parameters": [
          {
            "schema": {
              "maxLength": 120,
              "type": "string"
            },
            "in": "query",
            "name": "q",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "vegetable"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "fruit"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "herb"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "flower"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "cover_crop"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "type",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "familyId",
            "required": false
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "integer"
            },
            "in": "query",
            "name": "take",
            "required": false
          },
          {
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "integer"
            },
            "in": "query",
            "name": "skip",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "commonName",
                      "scientificName",
                      "slug",
                      "type",
                      "sun",
                      "water",
                      "feederType",
                      "spacingMm",
                      "rowSpacingMm",
                      "daysToMaturityMin",
                      "daysToMaturityMax",
                      "soilNotes",
                      "growingTips",
                      "commonMistakes",
                      "source",
                      "familyId",
                      "createdAt",
                      "updatedAt",
                      "deletedAt"
                    ],
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "commonName": {
                        "type": "string"
                      },
                      "scientificName": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "slug": {
                        "type": "string"
                      },
                      "type": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "vegetable"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "fruit"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "herb"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "flower"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "cover_crop"
                            ]
                          }
                        ]
                      },
                      "sun": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "water": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "feederType": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "spacingMm": {
                        "anyOf": [
                          {
                            "type": "integer"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "rowSpacingMm": {
                        "anyOf": [
                          {
                            "type": "integer"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "daysToMaturityMin": {
                        "anyOf": [
                          {
                            "type": "integer"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "daysToMaturityMax": {
                        "anyOf": [
                          {
                            "type": "integer"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "soilNotes": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "growingTips": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "commonMistakes": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "source": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "familyId": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "deletedAt": {
                        "anyOf": [
                          {
                            "type": "string",
                            "format": "date-time"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Standard error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "Standard error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "404": {
            "description": "Standard error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
    "/api/plants/{id}": {
      "get": {
        "summary": "Get a plant with its companion relationships",
        "tags": [
          "plants"
        ],
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "type": "object",
                      "required": [
                        "id",
                        "commonName",
                        "scientificName",
                        "slug",
                        "type",
                        "sun",
                        "water",
                        "feederType",
                        "spacingMm",
                        "rowSpacingMm",
                        "daysToMaturityMin",
                        "daysToMaturityMax",
                        "soilNotes",
                        "growingTips",
                        "commonMistakes",
                        "source",
                        "familyId",
                        "createdAt",
                        "updatedAt",
                        "deletedAt"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "commonName": {
                          "type": "string"
                        },
                        "scientificName": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "slug": {
                          "type": "string"
                        },
                        "type": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "vegetable"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "fruit"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "herb"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "flower"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "cover_crop"
                              ]
                            }
                          ]
                        },
                        "sun": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "water": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "feederType": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "spacingMm": {
                          "anyOf": [
                            {
                              "type": "integer"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "rowSpacingMm": {
                          "anyOf": [
                            {
                              "type": "integer"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "daysToMaturityMin": {
                          "anyOf": [
                            {
                              "type": "integer"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "daysToMaturityMax": {
                          "anyOf": [
                            {
                              "type": "integer"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "soilNotes": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "growingTips": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "commonMistakes": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "source": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "familyId": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "deletedAt": {
                          "anyOf": [
                            {
                              "type": "string",
                              "format": "date-time"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "companions"
                      ],
                      "properties": {
                        "companions": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": [
                              "plantId",
                              "relation",
                              "reason"
                            ],
                            "properties": {
                              "plantId": {
                                "type": "string"
                              },
                              "relation": {
                                "type": "string"
                              },
                              "reason": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Standard error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "Standard error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "404": {
            "description": "Standard error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
}